// packages/esm-patient-registration-app/src/patient-registration/patient-registration.types.ts
export interface PatientIdentifierValue {
  identifierUuid?: string;
  identifierTypeUuid: string;
  initialValue: string;
  identifierValue: string;
  identifierName: string;
  selectedSource: IdentifierSource;
  autoGeneration?: boolean;
  preferred: boolean;
  required: boolean;
}

/**
 * Extends the `FetchedPatientIdentifierType` with aggregated data.
 */
export interface PatientIdentifierType extends FetchedPatientIdentifierType {
  identifierSources: Array<IdentifierSource>;
  autoGenerationSource?: IdentifierSource;
  checked?: boolean;
  source?: IdentifierSource;
}

export interface IdentifierSource {
  uuid: string;
  name: string;
  autoGenerationOption?: IdentifierSourceAutoGenerationOption;
}
// ...
export interface RelationshipValue {
  relatedPersonName?: string;
  relatedPersonUuid: string;
  relation?: string;
  relationshipType: string;
  /**
   * Defines the action to be taken on the existing relationship
   * @kind ADD -> adds a new relationship
   * @kind UPDATE -> updates an existing relationship
   * @kind DELETE -> deletes an existing relationship
   * @kind undefined -> no operation on the existing relationship
   */
  action?: 'ADD' | 'UPDATE' | 'DELETE';
  /**
   * Value kept for restoring initial relationshipType value
   */
  initialrelationshipTypeValue?: string;
  uuid?: string;
}

export interface FormValues {
  additionalFamilyName: string;
  additionalGivenName: string;
  additionalMiddleName: string;
  addNameInLocalLanguage: boolean;
  address: {
    [addressField: string]: string;
  };
  attributes?: {
    [attributeTypeUuid: string]: string;
  };
  birthdate: Date | string;
  birthdateEstimated: boolean;
  deathCause: string;
  deathDate: string | Date;
  deathTime: string;
  deathTimeFormat: 'AM' | 'PM';
  familyName: string;
  gender: string;
  givenName: string;
  identifiers: {
    [identifierFieldName: string]: PatientIdentifierValue;
  };
  isDead: boolean;
  middleName: string;
  monthsEstimated: number;
  nonCodedCauseOfDeath: string;
  obs?: {
    [conceptUuid: string]: string;
  };
  patientUuid: string;
  relationships: Array<RelationshipValue>;
  telephoneNumber: string;
  yearsEstimated: number;
}

export interface PatientUuidMapType {
  additionalNameUuid?: string;
  preferredNameUuid?: string;
  preferredAddressUuid?: string;
}

export interface CapturePhotoProps {
  imageData: string;
  dateTime: string;
}